Moved native_geobox_ingested test
