+++ tests/test_testutils.py
@@ -56,6 +56,26 @@ def test_mk_sample_xr():
     assert mk_sample_xr_dataset(crs=None).geobox is None
 
 
+def test_native_geobox_ingested():
+    from datacube.testutils.io import native_geobox
+    from datacube.testutils.geom import AlbersGS
+
+    gbox = AlbersGS.tile_geobox((15, -40))
+    ds = mk_sample_dataset([dict(name='a')],
+                           geobox=gbox,
+                           product_opts=dict(with_grid_spec=True))
+
+    assert native_geobox(ds) == gbox
+
+    # check that dataset covering several tiles is detected as invalid
+    ds = mk_sample_dataset([dict(name='a')],
+                           geobox=gbox.buffered(10, 10),
+                           product_opts=dict(with_grid_spec=True))
+
+    with pytest.raises(ValueError):
+        native_geobox(ds)
+
+
 def test_native_geobox_eo3(eo3_dataset_s2):
     ds = eo3_dataset_s2
     assert ds.crs is not None
